requirements.txt
attrs==25.1.0
av==14.2.0
blinker==1.9.0
click==8.1.8
dlib==19.24.6
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
imutils==0.5.4
itsdangerous==2.2.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
libarchive-c==5.1
Mako==1.3.9
MarkupSafe==3.0.2
numpy==2.2.3
opencv-python==4.11.0.86
picamera2==0.3.25
pidng==4.0.9
piexif==1.1.3
pillow==11.1.0
python-prctl==1.8.1
referencing==0.36.2
rpds-py==0.23.1
RPi.GPIO==0.7.1
simplejpeg==1.8.2
SQLAlchemy==2.0.38
tqdm==4.67.1
typing_extensions==4.12.2
v4l2-python3==0.3.5
Werkzeug==3.1.3
